ui: split Bookmarks — project-wide admin moves to left sidebar, right sidebar shows active note only
The right-sidebar Bookmarks panel was showing the whole project tree
with rename / new-folder / drag-reorder / search affordances — that's
an administrative surface, which belongs in the left sidebar next to
Notes / Sources / Tags / Tables. The right sidebar should stay focused
on the active note.

- Move the existing full Bookmarks panel up to
  `components/BookmarksPanel.svelte` and add a 'Bookmarks' tab to the
  left Sidebar so the project-wide view is still one click away.
- Replace `right-sidebar/BookmarksPanel.svelte` with a minimal view:
  flat list of bookmarks for the active note, click to open, hover
  reveals a small delete affordance, "No bookmarks for this note"
  empty state. No folder/rename/drag UI.
- Drop `onShowPrompt` from RightSidebar's props — the minimal panel
  doesn't need to prompt for anything, and nothing else inside
  RightSidebar used it.
